+++ firmware/controllers/engine_controller.h
@@ -8,6 +8,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 // todo: huh we also have validateConfiguration()?!
 bool validateConfig();
 char* getPinNameByAdcChannel(const char* msg, adc_channel_e hwChannel, char* buffer);
@@ -20,3 +22,11 @@ void initDataStructures();
 void slowStartStopButtonCallback();
 
 void doPeriodicSlowCallback();
+
+struct SlowCallbackModuleDebug {
+	uint32_t totalUs = 0;
+	uint32_t topIndex[3] = {};
+	uint32_t topUs[3] = {};
+};
+
+SlowCallbackModuleDebug getSlowCallbackModuleDebug();