+++ java_console/models/src/main/java/com/rusefi/tracing/EnumNames.java
@@ -55,6 +55,11 @@ public class EnumNames {
 	"VvtHandleShaftSignal",
 	"WifiSpi",
 	"WifiHandleEvents",
-	"Idle"
+	"Idle",
+	"SlowAdcGap",
+	"SlowCallbackSensors",
+	"SlowCallbackModules",
+	"SlowCallbackTle8888",
+	"SlowCallbackBaro"
 	};
 }